+++ src/Type/EnumType/ServerEnum.php
@@ -13,7 +13,6 @@ class ServerEnum extends BaseType
 {
     const TYPE_RSYNC = 'rsync';
     const TYPE_EB = 'eb';
-    const TYPE_EC2 = 'ec2';
     const TYPE_S3 = 's3';
     const TYPE_CD = 'cd';
 
@@ -34,7 +33,6 @@ class ServerEnum extends BaseType
         return [
             self::TYPE_RSYNC,
             self::TYPE_EB,
-            self::TYPE_EC2,
             self::TYPE_S3,
             self::TYPE_CD,
         ];